Refactor button to new file
